Extract StorePickerErrorType into its own file
