+++ WooCommerce/Classes/Authentication/Epilogue/StorePickerError.swift
@@ -1,6 +1,5 @@
 import SwiftUI
 import SafariServices
-import enum NetworkingCore.DotcomError
 
 /// Hosting controller wrapper for `StorePickerError`
 ///
@@ -48,47 +47,6 @@ final class StorePickerErrorHostingController: UIHostingController<StorePickerEr
         rootView.troubleshootingAction = troubleshootingAction
         rootView.contactSupportAction = contactSupportAction
         rootView.dismissAction = dismissAction
-    }
-}
-
-/// Categorizes the error shown in the store picker error screen.
-/// Extracted as a standalone type for testability.
-///
-enum StorePickerErrorType: Equatable {
-    /// The remote site returned a server error (HTTP 500) when verifying user permissions.
-    case serverError
-    /// Any other error (network issues, unknown errors, etc.)
-    case generic
-
-    /// Creates an error type from the underlying error returned by the role eligibility check.
-    ///
-    static func from(_ error: Error) -> StorePickerErrorType {
-        if let dotcomError = error as? DotcomError,
-           case .requestFailed = dotcomError {
-            return .serverError
-        }
-        return .generic
-    }
-
-    var bodyText: String {
-        switch self {
-        case .serverError:
-            return Localization.serverErrorBody
-        case .generic:
-            return Localization.genericBody
-        }
-    }
-
-    private enum Localization {
-        static let serverErrorBody = NSLocalizedString(
-            "storePickerError.serverErrorBody",
-            value: "There was a problem verifying your permissions. " +
-                "Please contact your hosting provider for assistance, or reach out to us.",
-            comment: "Body text for the store picker error screen when the site returns a server error (HTTP 500)")
-        static let genericBody = NSLocalizedString(
-            "storePickerError.genericBody",
-            value: "Please try again or reach out to us and we'll be happy to assist you!",
-            comment: "Body text for the default store picker error screen")
     }
 }
 
@@ -235,7 +193,7 @@ struct StorePickerError_Preview: PreviewProvider {
 
         VStack {
             StorePickerError(errorType: .serverError,
-                             technicalDetails: String(describing: DotcomError.requestFailed()))
+                             technicalDetails: "requestFailed: http_request_failed")
         }
         .padding()
         .background(Color.gray)
